:truck: move `DeviceLike` hierarchy into a module
